Refactor photo map marker handling into controller
